avances en pruebas unitarias de DAO

+++ src/main/java/com/cubiktimer/modelo/dao/AmigoDAO.java
@@ -12,6 +12,7 @@ import org.apache.log4j.Logger;
 
 import com.cubiktimer.config.CubikTimerDataSource;
 import com.cubiktimer.modelo.dto.AmigoDTO;
+import com.cubiktimer.util.Constantes;
 import com.mysql.jdbc.Statement;
 
 public class AmigoDAO extends DAO<Integer, AmigoDTO> implements Serializable {
@@ -19,11 +20,17 @@ public class AmigoDAO extends DAO<Integer, AmigoDTO> implements Serializable {
 	private static final long serialVersionUID = 1L;
 	private static final Logger log = Logger.getLogger(AmigoDAO.class);
 
+	public AmigoDAO() {
+		super(Constantes.TABLA_AMIGOS);
+	}
+
 	@Override
 	public int create(AmigoDTO dto) {
+		if (dto == null) {
+			return 0;
+		}
 		log.trace("inicio create");
 		int retorno = 0;
-
 		String sql = "INSERT INTO amigos VALUES (?,?,?,?)";
 		try (Connection con = CubikTimerDataSource.getConnection();
 				PreparedStatement ps = con.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
@@ -40,21 +47,21 @@ public class AmigoDAO extends DAO<Integer, AmigoDTO> implements Serializable {
 			} finally {
 				rs.close();
 			}
-
 			log.trace("fin create");
 			return retorno;
 		} catch (Exception e) {
 			log.warn(e.getMessage());
-
 		}
 		log.trace("fin create");
 		return 0;
 	}
 
 	public int update(AmigoDTO dto) {
+		if (dto == null) {
+			return 0;
+		}
 		log.trace("inicio update");
 		int retorno = 0;
-
 		String sql = "UPDATE amigos SET id_usuario = ?, id_amigo = ?, estado = ? WHERE id_amistad = ?";
 		try (Connection con = CubikTimerDataSource.getConnection(); PreparedStatement ps = con.prepareStatement(sql)) {
 			ps.setObject(1, dto.getIdUsuario());
@@ -63,18 +70,19 @@ public class AmigoDAO extends DAO<Integer, AmigoDTO> implements Serializable {
 			ps.setObject(4, dto.getIdAmistad());
 			ps.executeUpdate();
 			retorno = dto.getIdUsuario();
-
 			log.trace("fin update");
 			return retorno;
 		} catch (Exception e) {
 			log.warn(e.getMessage());
-
 		}
 		log.trace("fin update");
 		return 0;
 	}
 
 	public int merge(AmigoDTO dto) {
+		if (dto == null) {
+			return 0;
+		}
 		if (dto.getIdAmistad() == null
 				&& consultarAmigosPorIdUsuarioYIdAmigo(dto.getIdUsuario(), dto.getIdAmigo()).isEmpty()) {
 			return create(dto);
@@ -92,20 +100,20 @@ public class AmigoDAO extends DAO<Integer, AmigoDTO> implements Serializable {
 	}
 
 	public void delete(AmigoDTO dto) {
+		if (dto == null) {
+			return;
+		}
 		log.trace("inicio delete");
 		if (dto.getIdAmistad() == null && sonAmigos(dto.getIdUsuario(), dto.getIdAmigo())) {
 			dto.setIdAmistad(
 					consultarAmigosPorIdUsuarioYIdAmigo(dto.getIdUsuario(), dto.getIdAmigo()).get(0).getIdAmistad());
 		}
-
 		String sql = "DELETE FROM amigos WHERE id_amistad = ?";
 		try (Connection con = CubikTimerDataSource.getConnection(); PreparedStatement ps = con.prepareStatement(sql)) {
 			ps.setObject(1, dto.getIdAmistad());
 			ps.executeUpdate();
-
 		} catch (Exception e) {
 			log.warn(e.getMessage());
-
 		}
 		log.trace("fin delete");
 	}
@@ -113,7 +121,6 @@ public class AmigoDAO extends DAO<Integer, AmigoDTO> implements Serializable {
 	public List<AmigoDTO> consultarAmigosPorIdUsuarioYIdAmigo(Integer idUsuario, Integer idAmigo) {
 		log.trace("inicio consultarAmigosPorIdUsuarioYIdAmigo");
 		List<AmigoDTO> lista = new ArrayList<>();
-
 		String sql = "SELECT * FROM amigos a WHERE a.id_usuario = ? AND a.id_amigo = ?";
 		try (Connection con = CubikTimerDataSource.getConnection(); PreparedStatement ps = con.prepareStatement(sql)) {
 			ps.setInt(1, idUsuario);
@@ -121,34 +128,25 @@ public class AmigoDAO extends DAO<Integer, AmigoDTO> implements Serializable {
 			lista = findList(ps);
 		} catch (SQLException e) {
 			log.warn(e.getMessage());
-		} finally {
-
 		}
 		log.trace("fin consultarAmigosPorIdUsuarioYIdAmigo");
 		return lista;
 	}
 
-	public List<AmigoDTO> findList(PreparedStatement ps) {
-		log.trace("inicio findList");
+	public List<AmigoDTO> setList(ResultSet rs) throws SQLException {
 		List<AmigoDTO> lista = new ArrayList<>();
-		try {
-			ResultSet rs = ps.executeQuery();
-			try {
-				while (rs.next()) {
-					AmigoDTO a = new AmigoDTO();
-					a.setIdAmistad(rs.getInt("id_amistad"));
-					a.setIdUsuario(rs.getInt("id_usuario"));
-					a.setIdAmigo(rs.getInt("id_amigo"));
-					a.setEstado(rs.getInt("estado"));
-					lista.add(a);
-				}
-			} finally {
-				rs.close();
-			}
-		} catch (SQLException sqle) {
-			log.warn(sqle.getMessage());
+		if (rs == null) {
+			return lista;
 		}
-		log.trace("fin findList");
+		while (rs.next()) {
+			AmigoDTO a = new AmigoDTO();
+			a.setIdAmistad(rs.getInt("id_amistad"));
+			a.setIdUsuario(rs.getInt("id_usuario"));
+			a.setIdAmigo(rs.getInt("id_amigo"));
+			a.setEstado(rs.getInt("estado"));
+			lista.add(a);
+		}
+		rs.close();
 		return lista;
 	}
